Added clone project dialog, edit project info dialog
